Remove legacy basic auth support
It was doing a `lookup` directly in the `labels.j2` template.

Despite this being in an `if` that shouldn't have triggered,
the `lookup` was being invoked and was erroring out due to being unable
to find the file.

This legacy auth method is not necessary anymore anyway, so it's better
to get rid of it.